requirements.txt
pyzmq>=18.1.1,<19.0.0
paho-mqtt>=1.5.0,<2.0.0
requests>=2.22.0,<3.0.0
python-dotenv

celery>=4.1.1
redis>=3.3.7

# From rfid_sync
pymongo>=3.10.1,<4.0.0

psutil>=5.7.0<6.0.0

python-dateutil==2.8.1
python-imap==1.0.0
cryptography>=2.9.2,<3.0.0
PyJWT>=1.6.4,<2.0.0
